Something is broken in the battle setup

+++ Dragoon Modifier/MemoryController/Battle/MonsterAddress.cs
@@ -17,7 +17,7 @@ namespace Dragoon_Modifier.MemoryController {
 
         public byte Element { get { return Emulator.ReadByte(_baseAddress + 0x6A); } set { Emulator.WriteByte(_baseAddress + 0x6A, value); } }
         public byte Display_Element { get { return Emulator.ReadByte(_baseAddress + 0x14); } set { Emulator.WriteByte(_baseAddress + 0x14, value); } }
-        public byte SigStone { get { return (byte) ((Emulator.ReadByte(_pShieldMshieldSigStone) >> 2) & 3); } set { Emulator.WriteByte(_pShieldMshieldSigStone, Emulator.ReadByte(_pShieldMshieldSigStone) | (Math.Max(value, (byte) 3) << 2)); } }
+        public byte SigStone { get { return (byte) ((Emulator.ReadByte(_pShieldMshieldSigStone) >> 2) & 3); } set { Emulator.WriteByte(_pShieldMshieldSigStone, (byte) (Emulator.ReadByte(_pShieldMshieldSigStone) | (Math.Max(value, (byte) 3) << 2))); } }
         public byte ChargingSpirit { get { return Emulator.ReadByte(_pShieldMshieldSigStone + 0x2); } set { Emulator.WriteByte(_pShieldMshieldSigStone + 0x2, value); } }
         public ushort Expirience { get { return Emulator.ReadUShort(_rewardsAddress + UniqueIndex * 0x1A8); } set { Emulator.WriteUShort(_rewardsAddress + UniqueIndex * 0x1A8, value); } }
         public ushort Gold { get { return Emulator.ReadUShort(_rewardsAddress + 0x2 + UniqueIndex * 0x1A8); } set { Emulator.WriteUShort(_rewardsAddress + 0x2 + UniqueIndex * 0x1A8, value); } }
